OTel.API.Trace.Core.Attributes

Documentation

pattern TRACE_SCHEMA_URL :: SchemaURL

The URL of the OpenTelemetry schema for these keys and values.

pattern AWS_LAMBDA_INVOKED_ARN :: Key Text

The full invoked ARN as provided on the Context passed to the function (Lambda-Runtime-Invoked-Function-Arn header on the /runtime/invocation/next applicable).

Notes

This may be different from faas.id if an alias is involved.

pattern CLOUDEVENTS_EVENT_ID :: Key Text

The https://github.com/cloudevents/spec/blob/v1.0.2/cloudevents/spec.md#id uniquely identifies the event.

pattern CLOUDEVENTS_EVENT_SOURCE :: Key Text

The https://github.com/cloudevents/spec/blob/v1.0.2/cloudevents/spec.md#source-1 identifies the context in which an event happened.

pattern CLOUDEVENTS_EVENT_SPEC_VERSION :: Key Text

The https://github.com/cloudevents/spec/blob/v1.0.2/cloudevents/spec.md#specversion which the event uses.

pattern CLOUDEVENTS_EVENT_TYPE :: Key Text

The https://github.com/cloudevents/spec/blob/v1.0.2/cloudevents/spec.md#type contains a value describing the type of event related to the originating occurrence.

pattern CLOUDEVENTS_EVENT_SUBJECT :: Key Text

The https://github.com/cloudevents/spec/blob/v1.0.2/cloudevents/spec.md#subject of the event in the context of the event producer (identified by source).

pattern OPENTRACING_REF_TYPE :: Key Text

Parent-child Reference type

Notes

The causal relationship between a child Span and a parent Span.

pattern DB_SYSTEM :: Key Text

An identifier for the database management system (DBMS) product being used. See below for a list of well-known identifiers.

pattern DB_CONNECTION_STRING :: Key Text

The connection string used to connect to the database. It is recommended to remove embedded credentials.

pattern DB_USER :: Key Text

Username for accessing the database.

pattern DB_JDBC_DRIVER_CLASSNAME :: Key Text

The fully-qualified class name of the https://docs.oracle.com/javase/8/docs/technotes/guides/jdbc/ driver used to connect.

pattern DB_NAME :: Key Text

This attribute is used to report the name of the database being accessed. For commands that switch the database, this should be set to the target database (even if the command fails).

Notes

In some SQL databases, the database name to be used is called &quot;schema name&quot;. In case there are multiple layers that could be considered for database name (e.g. Oracle instance name and schema name), the database name to be used is the more specific layer (e.g. Oracle schema name).

pattern DB_STATEMENT :: Key Text

The database statement being executed.

Notes

The value may be sanitized to exclude sensitive information.

pattern DB_OPERATION :: Key Text

The name of the operation being executed, e.g. the https://docs.mongodb.com/manual/reference/command/#database-operations such as findAndModify, or the SQL keyword.

Notes

When setting this to an SQL keyword, it is not recommended to attempt any client-side parsing of db.statement just to get this property, but it should be set if the operation name is provided by the library being instrumented. If the SQL statement has an ambiguous operation, or performs more than one operation, this value may be omitted.

pattern DB_MSSQL_INSTANCE_NAME :: Key Text

The Microsoft SQL Server https://docs.microsoft.com/en-us/sql/connect/jdbc/building-the-connection-url?view=sql-server-ver15 connecting to. This name is used to determine the port of a named instance.

Notes

If setting a db.mssql.instance_name, net.peer.port is no longer required (but still recommended if non-standard).

pattern DB_CASSANDRA_PAGE_SIZE :: Key Int64

The fetch size used for paging, i.e. how many rows will be returned at once.

pattern DB_CASSANDRA_CONSISTENCY_LEVEL :: Key Text

The consistency level of the query. Based on consistency values from https://docs.datastax.com/en/cassandra-oss/3.0/cassandra/dml/dmlConfigConsistency.html.

pattern DB_CASSANDRA_TABLE :: Key Text

The name of the primary table that the operation is acting upon, including the keyspace name (if applicable).

Notes

This mirrors the db.sql.table attribute but references cassandra rather than sql. It is not recommended to attempt any client-side parsing of db.statement just to get this property, but it should be set if it is provided by the library being instrumented. If the operation is acting upon an anonymous table, or more than one table, this value MUST NOT be set.

pattern DB_CASSANDRA_IDEMPOTENCE :: Key Bool

Whether or not the query is idempotent.

pattern DB_CASSANDRA_SPECULATIVE_EXECUTION_COUNT :: Key Int64

The number of times a query was speculatively executed. Not set or 0 if the query was not executed speculatively.

pattern DB_CASSANDRA_COORDINATOR_ID :: Key Text

The ID of the coordinating node for a query.

pattern DB_CASSANDRA_COORDINATOR_DC :: Key Text

The data center of the coordinating node for a query.

pattern DB_REDIS_DATABASE_INDEX :: Key Int64

The index of the database being accessed as used in the https://redis.io/commands/select, provided as an integer. To be used instead of the generic db.name attribute.

pattern DB_MONGODB_COLLECTION :: Key Text

The collection being accessed within the database stated in db.name.

pattern DB_SQL_TABLE :: Key Text

The name of the primary table that the operation is acting upon, including the database name (if applicable).

Notes

It is not recommended to attempt any client-side parsing of db.statement just to get this property, but it should be set if it is provided by the library being instrumented. If the operation is acting upon an anonymous table, or more than one table, this value MUST NOT be set.

pattern EXCEPTION_TYPE :: Key Text

The type of the exception (its fully-qualified class name, if applicable). The dynamic type of the exception should be preferred over the static type in languages that support it.

pattern EXCEPTION_MESSAGE :: Key Text

The exception message.

pattern EXCEPTION_STACKTRACE :: Key Text

A stacktrace as a string in the natural representation for the language runtime. The representation is to be determined and documented by each language SIG.

pattern EXCEPTION_ESCAPED :: Key Bool

SHOULD be set to true if the exception event is recorded at a point where it is known that the exception is escaping the scope of the span.

Notes

An exception is considered to have escaped (or left) the scope of a span, if that span is ended while the exception is still logically &quot;in flight&quot;. This may be actually &quot;in flight&quot; in some languages (e.g. if the exception is passed to a Context manager's exit method in Python) but will usually be caught at the point of recording the exception in most languages.It is usually not possible to determine at the point where an exception is thrown whether it will escape the scope of a span. However, it is trivial to know that an exception will escape, if one checks for an active exception just before ending the span, as done in the #recording-an-exception.It follows that an exception may still escape the scope of the span even if the exception.escaped attribute was not set or set to false, since the event might have been recorded at a time where it was not clear whether the exception will escape.

pattern FAAS_TRIGGER :: Key Text

Type of the trigger which caused this function execution.

Notes

For the server/consumer span on the incoming side, faas.trigger MUST be set.Clients invoking FaaS instances usually cannot set faas.trigger, since they would typically need to look in the payload to determine the event type. If clients set it, it should be the same as the trigger that corresponding incoming would have (i.e., this has nothing to do with the underlying transport used to make the API call to invoke the lambda, which is often HTTP).

pattern FAAS_EXECUTION :: Key Text

The execution ID of the current function execution.

pattern FAAS_DOCUMENT_COLLECTION :: Key Text

The name of the source on which the triggering operation was performed. For example, in Cloud Storage or S3 corresponds to the bucket name, and in Cosmos DB to the database name.

pattern FAAS_DOCUMENT_OPERATION :: Key Text

Describes the type of the operation that was performed on the data.

pattern FAAS_DOCUMENT_TIME :: Key Text

A string containing the time when the data was accessed in the https://www.iso.org/iso-8601-date-and-time-format.html format expressed in https://www.w3.org/TR/NOTE-datetime.

pattern FAAS_DOCUMENT_NAME :: Key Text

The document name/table subjected to the operation. For example, in Cloud Storage or S3 is the name of the file, and in Cosmos DB the table name.

pattern FAAS_TIME :: Key Text

A string containing the function invocation time in the https://www.iso.org/iso-8601-date-and-time-format.html format expressed in https://www.w3.org/TR/NOTE-datetime.

pattern FAAS_CRON :: Key Text

A string containing the schedule period as https://docs.oracle.com/cd/E12058_01/doc/doc.1014/e12030/cron_expressions.htm.

pattern FAAS_COLDSTART :: Key Bool

A boolean that is true if the serverless function is executed for the first time (aka cold-start).

pattern FAAS_INVOKED_NAME :: Key Text

The name of the invoked function.

Notes

SHOULD be equal to the faas.name resource attribute of the invoked function.

pattern FAAS_INVOKED_PROVIDER :: Key Text

The cloud provider of the invoked function.

Notes

SHOULD be equal to the cloud.provider resource attribute of the invoked function.

pattern FAAS_INVOKED_REGION :: Key Text

The cloud region of the invoked function.

Notes

SHOULD be equal to the cloud.region resource attribute of the invoked function.

pattern NET_TRANSPORT :: Key Text

Transport protocol used. See note below.

pattern NET_PEER_IP :: Key Text

Remote address of the peer (dotted decimal for IPv4 or https://tools.ietf.org/html/rfc5952 for IPv6)

pattern NET_PEER_PORT :: Key Int64

Remote port number.

pattern NET_PEER_NAME :: Key Text

Remote hostname or similar, see note below.

Notes

net.peer.name SHOULD NOT be set if capturing it would require an extra DNS lookup.

pattern NET_HOST_IP :: Key Text

Like net.peer.ip but for the host IP. Useful in case of a multi-IP host.

pattern NET_HOST_PORT :: Key Int64

Like net.peer.port but for the host port.

pattern NET_HOST_NAME :: Key Text

Local hostname or similar, see note below.

pattern NET_HOST_CONNECTION_TYPE :: Key Text

The internet connection type currently being used by the host.

pattern NET_HOST_CONNECTION_SUBTYPE :: Key Text

This describes more details regarding the connection.type. It may be the type of cell technology connection, but it could be used for describing details about a wifi connection.

pattern NET_HOST_CARRIER_NAME :: Key Text

The name of the mobile carrier.

pattern NET_HOST_CARRIER_MCC :: Key Text

The mobile carrier country code.

pattern NET_HOST_CARRIER_MNC :: Key Text

The mobile carrier network code.

pattern NET_HOST_CARRIER_ICC :: Key Text

The ISO 3166-1 alpha-2 2-character country code associated with the mobile carrier network.

pattern PEER_SERVICE :: Key Text

The ../../resource/semantic_conventions/README.md#service of the remote service. SHOULD be equal to the actual service.name resource attribute of the remote service if any.

pattern ENDUSER_ID :: Key Text

Username or client_id extracted from the access token or https://tools.ietf.org/html/rfc7235#section-4.2 header in the inbound request from outside the system.

pattern ENDUSER_ROLE :: Key Text

Actual/assumed role the client is making the request under extracted from token or application security context.

pattern ENDUSER_SCOPE :: Key Text

Scopes or granted authorities the client currently possesses extracted from token or application security context. The value would come from the scope associated with an https://tools.ietf.org/html/rfc6749#section-3.3 or an attribute value in a http://docs.oasis-open.org/security/saml/Post2.0/sstc-saml-tech-overview-2.0.html.

pattern THREAD_ID :: Key Int64

Current &quot;managed&quot; thread ID (as opposed to OS thread ID).

pattern THREAD_NAME :: Key Text

Current thread name.

pattern CODE_FUNCTION :: Key Text

The method or function name, or equivalent (usually rightmost part of the code unit's name).

pattern CODE_NAMESPACE :: Key Text

The &quot;namespace&quot; within which code.function is defined. Usually the qualified class or module name, such that code.namespace + some separator + code.function form a unique identifier for the code unit.

pattern CODE_FILEPATH :: Key Text

The source code file name that identifies the code unit as uniquely as possible (preferably an absolute file path).

pattern CODE_LINENO :: Key Int64

The line number in code.filepath best representing the operation. It SHOULD point within the code unit named in code.function.

pattern HTTP_METHOD :: Key Text

HTTP request method.

pattern HTTP_URL :: Key Text

Full HTTP request URL in the form scheme://host[:port]/path?query[#fragment]. Usually the fragment is not transmitted over HTTP, but if it is known, it should be included nevertheless.

Notes

http.url MUST NOT contain credentials passed via URL in form of https://username:passwordwww.example.com/. In such case the attribute's value should be https://www.example.com/@.

pattern HTTP_TARGET :: Key Text

The full request target as passed in a HTTP request line or equivalent.

pattern HTTP_HOST :: Key Text

The value of the https://tools.ietf.org/html/rfc7230#section-5.4. An empty Host header should also be reported, see note.

Notes

When the header is present but empty the attribute SHOULD be set to the empty string. Note that this is a valid situation that is expected in certain cases, according the aforementioned https://tools.ietf.org/html/rfc7230#section-5.4. When the header is not set the attribute MUST NOT be set.

pattern HTTP_SCHEME :: Key Text

The URI scheme identifying the used protocol.

pattern HTTP_STATUS_CODE :: Key Int64

https://tools.ietf.org/html/rfc7231#section-6.

pattern HTTP_FLAVOR :: Key Text

Kind of HTTP protocol used.

Notes

If net.transport is not specified, it can be assumed to be IP.TCP except if http.flavor is QUIC, in which case IP.UDP is assumed.

pattern HTTP_USER_AGENT :: Key Text

Value of the https://tools.ietf.org/html/rfc7231#section-5.5.3 header sent by the client.

pattern HTTP_REQUEST_CONTENT_LENGTH :: Key Int64

The size of the request payload body in bytes. This is the number of bytes transferred excluding headers and is often, but not always, present as the https://tools.ietf.org/html/rfc7230#section-3.3.2 header. For requests using transport encoding, this should be the compressed size.

pattern HTTP_REQUEST_CONTENT_LENGTH_UNCOMPRESSED :: Key Int64

The size of the uncompressed request payload body after transport decoding. Not set if transport encoding not used.

pattern HTTP_RESPONSE_CONTENT_LENGTH :: Key Int64

The size of the response payload body in bytes. This is the number of bytes transferred excluding headers and is often, but not always, present as the https://tools.ietf.org/html/rfc7230#section-3.3.2 header. For requests using transport encoding, this should be the compressed size.

pattern HTTP_RESPONSE_CONTENT_LENGTH_UNCOMPRESSED :: Key Int64

The size of the uncompressed response payload body after transport decoding. Not set if transport encoding not used.

pattern HTTP_RETRY_COUNT :: Key Int64

The ordinal number of request re-sending attempt.

pattern HTTP_SERVER_NAME :: Key Text

The primary server name of the matched virtual host. This should be obtained via configuration. If no such configuration can be obtained, this attribute MUST NOT be set ( net.host.name should be used instead).

Notes

http.url is usually not readily available on the server side but would have to be assembled in a cumbersome and sometimes lossy process from other information (see e.g. open-telemetryopentelemetry-pythonpull/148). It is thus preferred to supply the raw data that is available.

pattern HTTP_ROUTE :: Key Text

The matched route (path template).

pattern HTTP_CLIENT_IP :: Key Text

The IP address of the original client behind all proxies, if known (e.g. from https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/X-Forwarded-For).

Notes

This is not necessarily the same as net.peer.ip, which would identify the network-level peer, which may be a proxy.This attribute should be set when a source of information different from the one used for net.peer.ip, is available even if that other source just confirms the same value as net.peer.ip. Rationale: For net.peer.ip, one typically does not know if it comes from a proxy, reverse proxy, or the actual client. Setting http.client_ip when it's the same as net.peer.ip means that one is at least somewhat confident that the address is not that of the closest proxy.

pattern AWS_DYNAMODB_TABLE_NAMES :: Key (AttrVals Text)

The keys in the RequestItems object field.

pattern AWS_DYNAMODB_CONSUMED_CAPACITY :: Key (AttrVals Text)

The JSON-serialized value of each item in the ConsumedCapacity response field.

pattern AWS_DYNAMODB_ITEM_COLLECTION_METRICS :: Key Text

The JSON-serialized value of the ItemCollectionMetrics response field.

pattern AWS_DYNAMODB_PROVISIONED_READ_CAPACITY :: Key Double

The value of the ProvisionedThroughput.ReadCapacityUnits request parameter.

pattern AWS_DYNAMODB_PROVISIONED_WRITE_CAPACITY :: Key Double

The value of the ProvisionedThroughput.WriteCapacityUnits request parameter.

pattern AWS_DYNAMODB_CONSISTENT_READ :: Key Bool

The value of the ConsistentRead request parameter.

pattern AWS_DYNAMODB_PROJECTION :: Key Text

The value of the ProjectionExpression request parameter.

pattern AWS_DYNAMODB_LIMIT :: Key Int64

The value of the Limit request parameter.

pattern AWS_DYNAMODB_ATTRIBUTES_TO_GET :: Key (AttrVals Text)

The value of the AttributesToGet request parameter.

pattern AWS_DYNAMODB_INDEX_NAME :: Key Text

The value of the IndexName request parameter.

pattern AWS_DYNAMODB_SELECT :: Key Text

The value of the Select request parameter.

pattern AWS_DYNAMODB_GLOBAL_SECONDARY_INDEXES :: Key (AttrVals Text)

The JSON-serialized value of each item of the GlobalSecondaryIndexes request field

pattern AWS_DYNAMODB_LOCAL_SECONDARY_INDEXES :: Key (AttrVals Text)

The JSON-serialized value of each item of the LocalSecondaryIndexes request field.

pattern AWS_DYNAMODB_EXCLUSIVE_START_TABLE :: Key Text

The value of the ExclusiveStartTableName request parameter.

pattern AWS_DYNAMODB_TABLE_COUNT :: Key Int64

The the number of items in the TableNames response parameter.

pattern AWS_DYNAMODB_SCAN_FORWARD :: Key Bool

The value of the ScanIndexForward request parameter.

pattern AWS_DYNAMODB_SEGMENT :: Key Int64

The value of the Segment request parameter.

pattern AWS_DYNAMODB_TOTAL_SEGMENTS :: Key Int64

The value of the TotalSegments request parameter.

pattern AWS_DYNAMODB_COUNT :: Key Int64

The value of the Count response parameter.

pattern AWS_DYNAMODB_SCANNED_COUNT :: Key Int64

The value of the ScannedCount response parameter.

pattern AWS_DYNAMODB_ATTRIBUTE_DEFINITIONS :: Key (AttrVals Text)

The JSON-serialized value of each item in the AttributeDefinitions request field.

pattern AWS_DYNAMODB_GLOBAL_SECONDARY_INDEX_UPDATES :: Key (AttrVals Text)

The JSON-serialized value of each item in the the GlobalSecondaryIndexUpdates request field.

pattern MESSAGING_SYSTEM :: Key Text

A string identifying the messaging system.

pattern MESSAGING_DESTINATION :: Key Text

The message destination name. This might be equal to the span name but is required nevertheless.

pattern MESSAGING_DESTINATION_KIND :: Key Text

The kind of message destination

pattern MESSAGING_TEMP_DESTINATION :: Key Bool

A boolean that is true if the message destination is temporary.

pattern MESSAGING_PROTOCOL :: Key Text

The name of the transport protocol.

pattern MESSAGING_PROTOCOL_VERSION :: Key Text

The version of the transport protocol.

pattern MESSAGING_URL :: Key Text

Connection string.

pattern MESSAGING_MESSAGE_ID :: Key Text

A value used by the messaging system as an identifier for the message, represented as a string.

pattern MESSAGING_CONVERSATION_ID :: Key Text

The #conversations identifying the conversation to which the message belongs, represented as a string. Sometimes called &quot;Correlation ID&quot;.

pattern MESSAGING_MESSAGE_PAYLOAD_SIZE_BYTES :: Key Int64

The (uncompressed) size of the message payload in bytes. Also use this attribute if it is unknown whether the compressed or uncompressed payload size is reported.

pattern MESSAGING_MESSAGE_PAYLOAD_COMPRESSED_SIZE_BYTES :: Key Int64

The compressed size of the message payload in bytes.

pattern MESSAGING_OPERATION :: Key Text

A string identifying the kind of message consumption as defined in the #operation-names section above. If the operation is &quot;send&quot;, this attribute MUST NOT be set, since the operation can be inferred from the span kind in that case.

pattern MESSAGING_CONSUMER_ID :: Key Text

The identifier for the consumer receiving a message. For Kafka, set it to {messaging.kafka.consumer_group} - {messaging.kafka.client_id}, if both are present, or only messaging.kafka.consumer_group. For brokers, such as RabbitMQ and Artemis, set it to the client_id of the client consuming the message.

pattern MESSAGING_RABBITMQ_ROUTING_KEY :: Key Text

RabbitMQ message routing key.

pattern MESSAGING_KAFKA_MESSAGE_KEY :: Key Text

Message keys in Kafka are used for grouping alike messages to ensure they're processed on the same partition. They differ from messaging.message_id in that they're not unique. If the key is null, the attribute MUST NOT be set.

Notes

If the key type is not string, it's string representation has to be supplied for the attribute. If the key has no unambiguous, canonical string form, don't include its value.

pattern MESSAGING_KAFKA_CONSUMER_GROUP :: Key Text

Name of the Kafka Consumer Group that is handling the message. Only applies to consumers, not producers.

pattern MESSAGING_KAFKA_CLIENT_ID :: Key Text

Client Id for the Consumer or Producer that is handling the message.

pattern MESSAGING_KAFKA_PARTITION :: Key Int64

Partition the message is sent to.

pattern MESSAGING_KAFKA_TOMBSTONE :: Key Bool

A boolean that is true if the message is a tombstone.

pattern MESSAGING_ROCKETMQ_NAMESPACE :: Key Text

Namespace of RocketMQ resources, resources in different namespaces are individual.

pattern MESSAGING_ROCKETMQ_CLIENT_GROUP :: Key Text

Name of the RocketMQ producer/consumer group that is handling the message. The client type is identified by the SpanKind.

pattern MESSAGING_ROCKETMQ_CLIENT_ID :: Key Text

The unique identifier for each client.

pattern MESSAGING_ROCKETMQ_MESSAGE_TYPE :: Key Text

Type of message.

pattern MESSAGING_ROCKETMQ_MESSAGE_TAG :: Key Text

The secondary classifier of message besides topic.

pattern MESSAGING_ROCKETMQ_MESSAGE_KEYS :: Key (AttrVals Text)

Key(s) of message, another way to mark message besides message id.

pattern MESSAGING_ROCKETMQ_CONSUMPTION_MODEL :: Key Text

Model of message consumption. This only applies to consumer spans.

pattern RPC_SYSTEM :: Key Text

A string identifying the remoting system. See below for a list of well-known identifiers.

pattern RPC_SERVICE :: Key Text

The full (logical) name of the service being called, including its package name, if applicable.

Notes

This is the logical name of the service from the RPC interface perspective, which can be different from the name of any implementing class. The code.namespace attribute may be used to store the latter (despite the attribute name, it may include a class name; e.g., class with method actually executing the call on the server side, RPC client stub class on the client side).

pattern RPC_METHOD :: Key Text

The name of the (logical) method being called, must be equal to the $method part in the span name.

Notes

This is the logical name of the method from the RPC interface perspective, which can be different from the name of any implementing method/function. The code.function attribute may be used to store the latter (e.g., method actually executing the call on the server side, RPC client stub method on the client side).

pattern RPC_GRPC_STATUS_CODE :: Key Int64

The https://github.com/grpc/grpc/blob/v1.33.2/doc/statuscodes.md of the gRPC request.

pattern RPC_JSONRPC_VERSION :: Key Text

Protocol version as in jsonrpc property of request/response. Since JSON-RPC 1.0 does not specify this, the value can be omitted.

pattern RPC_JSONRPC_REQUEST_ID :: Key Text

id property of request or response. Since protocol allows id to be int, string, null or missing (for notifications), value is expected to be cast to string for simplicity. Use empty string in case of null value. Omit entirely if this is a notification.

pattern RPC_JSONRPC_ERROR_CODE :: Key Int64

error.code property of response if it is an error response.

pattern RPC_JSONRPC_ERROR_MESSAGE :: Key Text

error.message property of response if it is an error response.

pattern MESSAGE_TYPE :: Key Text

Whether this is a received or sent message.

pattern MESSAGE_ID :: Key Int64

MUST be calculated as two different counters starting from 1 one for sent messages and one for received message.

Notes

This way we guarantee that the values will be consistent between different implementations.

pattern MESSAGE_COMPRESSED_SIZE :: Key Int64

Compressed size of the message in bytes.

pattern MESSAGE_UNCOMPRESSED_SIZE :: Key Int64

Uncompressed size of the message in bytes.

data OpentracingRefType

Constructors

OpentracingRefTypeChildOf
The parent Span depends on the child Span in some capacity.

OpentracingRefTypeFollowsFrom
The parent Span does not depend in any way on the result of the child Span.

data DbSystem

Constructors

DbSystemOtherSql
Some other SQL database. Fallback only. See notes.

DbSystemMssql
Microsoft SQL Server.

DbSystemMysql
MySQL.

DbSystemOracle
Oracle Database.

DbSystemDb2
IBM Db2.

DbSystemPostgresql
PostgreSQL.

DbSystemRedshift
Amazon Redshift.

DbSystemHive
Apache Hive.

DbSystemCloudscape
Cloudscape.

DbSystemHsqldb
HyperSQL DataBase.

DbSystemProgress
Progress Database.

DbSystemMaxdb
SAP MaxDB.

DbSystemHanadb
SAP HANA.

DbSystemIngres
Ingres.

DbSystemFirstsql
FirstSQL.

DbSystemEdb
EnterpriseDB.

DbSystemCache
InterSystems Caché.

DbSystemAdabas
Adabas (Adaptable Database System).

DbSystemFirebird
Firebird.

DbSystemDerby
Apache Derby.

DbSystemFilemaker
FileMaker.

DbSystemInformix
Informix.

DbSystemInstantdb
InstantDB.

DbSystemInterbase
InterBase.

DbSystemMariadb
MariaDB.

DbSystemNetezza
Netezza.

DbSystemPervasive
Pervasive PSQL.

DbSystemPointbase
PointBase.

DbSystemSqlite
SQLite.

DbSystemSybase
Sybase.

DbSystemTeradata
Teradata.

DbSystemVertica
Vertica.

DbSystemH2
H2.

DbSystemColdfusion
ColdFusion IMQ.

DbSystemCassandra
Apache Cassandra.

DbSystemHbase
Apache HBase.

DbSystemMongodb
MongoDB.

DbSystemRedis
Redis.

DbSystemCouchbase
Couchbase.

DbSystemCouchdb
CouchDB.

DbSystemCosmosdb
Microsoft Azure Cosmos DB.

DbSystemDynamodb
Amazon DynamoDB.

DbSystemNeo4j
Neo4j.

DbSystemGeode
Apache Geode.

DbSystemElasticsearch
Elasticsearch.

DbSystemMemcached
Memcached.

DbSystemCockroachdb
CockroachDB.

data DbCassandraConsistencyLevel

Constructors

DbCassandraConsistencyLevelAll
all.

DbCassandraConsistencyLevelEachQuorum
each_quorum.

DbCassandraConsistencyLevelQuorum
quorum.

DbCassandraConsistencyLevelLocalQuorum
local_quorum.

DbCassandraConsistencyLevelOne
one.

DbCassandraConsistencyLevelTwo
two.

DbCassandraConsistencyLevelThree
three.

DbCassandraConsistencyLevelLocalOne
local_one.

DbCassandraConsistencyLevelAny
any.

DbCassandraConsistencyLevelSerial
serial.

DbCassandraConsistencyLevelLocalSerial
local_serial.

data FaasTrigger

Constructors

FaasTriggerDatasource
A response to some data source operation such as a database or filesystem read/write.

FaasTriggerHttp
To provide an answer to an inbound HTTP request.

FaasTriggerPubsub
A function is set to be executed when messages are sent to a messaging system.

FaasTriggerTimer
A function is scheduled to be executed regularly.

FaasTriggerOther
If none of the others apply.

data FaasDocumentOperation

Constructors

FaasDocumentOperationInsert
When a new object is created.

FaasDocumentOperationEdit
When an object is modified.

FaasDocumentOperationDelete
When an object is deleted.

data FaasInvokedProvider

Constructors

FaasInvokedProviderAlibabaCloud
Alibaba Cloud.

FaasInvokedProviderAws
Amazon Web Services.

FaasInvokedProviderAzure
Microsoft Azure.

FaasInvokedProviderGcp
Google Cloud Platform.

FaasInvokedProviderTencentCloud
Tencent Cloud.

data NetTransport

Constructors

NetTransportIpTcp
ip_tcp.

NetTransportIpUdp
ip_udp.

NetTransportIp
Another IP-based protocol.

NetTransportUnix
Unix Domain socket. See below.

NetTransportPipe
Named or anonymous pipe. See note below.

NetTransportInproc
In-process communication.

NetTransportOther
Something else (non IP-based).

data NetHostConnectionType

Constructors

NetHostConnectionTypeWifi
wifi.

NetHostConnectionTypeWired
wired.

NetHostConnectionTypeCell
cell.

NetHostConnectionTypeUnavailable
unavailable.

NetHostConnectionTypeUnknown
unknown.

data NetHostConnectionSubtype

Constructors

NetHostConnectionSubtypeGprs
GPRS.

NetHostConnectionSubtypeEdge
EDGE.

NetHostConnectionSubtypeUmts
UMTS.

NetHostConnectionSubtypeCdma
CDMA.

NetHostConnectionSubtypeEvdo0
EVDO Rel. 0.

NetHostConnectionSubtypeEvdoA
EVDO Rev. A.

NetHostConnectionSubtypeCdma20001xrtt
CDMA2000 1XRTT.

NetHostConnectionSubtypeHsdpa
HSDPA.

NetHostConnectionSubtypeHsupa
HSUPA.

NetHostConnectionSubtypeHspa
HSPA.

NetHostConnectionSubtypeIden
IDEN.

NetHostConnectionSubtypeEvdoB
EVDO Rev. B.

NetHostConnectionSubtypeLte
LTE.

NetHostConnectionSubtypeEhrpd
EHRPD.

NetHostConnectionSubtypeHspap
HSPAP.

NetHostConnectionSubtypeGsm
GSM.

NetHostConnectionSubtypeTdScdma
TD-SCDMA.

NetHostConnectionSubtypeIwlan
IWLAN.

NetHostConnectionSubtypeNr
5G NR (New Radio).

NetHostConnectionSubtypeNrnsa
5G NRNSA (New Radio Non-Standalone).

NetHostConnectionSubtypeLteCa
LTE CA.

data HttpFlavor

Constructors

HttpFlavorHttp10
HTTP/1.0.

HttpFlavorHttp11
HTTP/1.1.

HttpFlavorHttp20
HTTP/2.

HttpFlavorHttp30
HTTP/3.

HttpFlavorSpdy
SPDY protocol.

HttpFlavorQuic
QUIC protocol.

data MessagingDestinationKind

Constructors

MessagingDestinationKindQueue
A message sent to a queue.

MessagingDestinationKindTopic
A message sent to a topic.

data MessagingOperation

Constructors

MessagingOperationReceive
receive.

MessagingOperationProcess
process.

data MessagingRocketmqMessageType

Constructors

MessagingRocketmqMessageTypeNormal
Normal message.

MessagingRocketmqMessageTypeFifo
FIFO message.

MessagingRocketmqMessageTypeDelay
Delay message.

MessagingRocketmqMessageTypeTransaction
Transaction message.

data MessagingRocketmqConsumptionModel

Constructors

MessagingRocketmqConsumptionModelClustering
Clustering consumption model.

MessagingRocketmqConsumptionModelBroadcasting
Broadcasting consumption model.

data RpcSystem

Constructors

RpcSystemGrpc
gRPC.

RpcSystemJavaRmi
Java RMI.

RpcSystemDotnetWcf
.NET WCF.

RpcSystemApacheDubbo
Apache Dubbo.

data RpcGrpcStatusCode

Constructors

RpcGrpcStatusCodeOk
OK.

RpcGrpcStatusCodeCancelled
CANCELLED.

RpcGrpcStatusCodeUnknown
UNKNOWN.

RpcGrpcStatusCodeInvalidArgument
INVALID_ARGUMENT.

RpcGrpcStatusCodeDeadlineExceeded
DEADLINE_EXCEEDED.

RpcGrpcStatusCodeNotFound
NOT_FOUND.

RpcGrpcStatusCodeAlreadyExists
ALREADY_EXISTS.

RpcGrpcStatusCodePermissionDenied
PERMISSION_DENIED.

RpcGrpcStatusCodeResourceExhausted
RESOURCE_EXHAUSTED.

RpcGrpcStatusCodeFailedPrecondition
FAILED_PRECONDITION.

RpcGrpcStatusCodeAborted
ABORTED.

RpcGrpcStatusCodeOutOfRange
OUT_OF_RANGE.

RpcGrpcStatusCodeUnimplemented
UNIMPLEMENTED.

RpcGrpcStatusCodeInternal
INTERNAL.

RpcGrpcStatusCodeUnavailable
UNAVAILABLE.

RpcGrpcStatusCodeDataLoss
DATA_LOSS.

RpcGrpcStatusCodeUnauthenticated
UNAUTHENTICATED.

data MessageType

Constructors

MessageTypeSent
sent.

MessageTypeReceived
received.

pattern DB_CASSANDRA_KEYSPACE :: Key Text

Deprecated: This item has been removed as of 1.8.0 of the semantic conventions. Please use DB_NAME instead.

The name of the keyspace being accessed.

Deprecated: This item has been removed as of 1.8.0 of the semantic conventions. Please use DB_NAME instead.

pattern DB_HBASE_NAMESPACE :: Key Text

Deprecated: This item has been removed as of 1.8.0 of the semantic conventions. Please use DB_NAME instead.

The HBase namespace being accessed.

Deprecated: This item has been removed as of 1.8.0 of the semantic conventions. Please use DB_NAME instead.