OTel.API.Trace.Core.Internal

Disclaimer

In general, changes to this module will not be reflected in the library's version updates. Direct use of this module should be done with utmost care, otherwise invariants will easily be violated.

trace_ :: (MonadTracing m, HasCallStack) => SpanSpec -> m a -> m a

trace :: (MonadTracing m, HasCallStack) => SpanSpec -> (MutableSpan -> m a) -> m a

class Monad m => MonadTracing m where

Minimal complete definition

Nothing

Methods

traceCS :: CallStack -> SpanSpec -> (MutableSpan -> m a) -> m a

default traceCS :: (MonadTransControl t, MonadTracing n, m ~ t n) => CallStack -> SpanSpec -> (MutableSpan -> m a) -> m a

getSpanContext :: MutableSpan -> m SpanContext

default getSpanContext :: (MonadTrans t, MonadTracing n, m ~ t n) => MutableSpan -> m SpanContext

updateSpan :: MutableSpan -> UpdateSpanSpec -> m ()

default updateSpan :: (MonadTrans t, MonadTracing n, m ~ t n) => MutableSpan -> UpdateSpanSpec -> m ()

class (MonadTracing m, MonadIO m) => MonadTracingIO m where

Minimal complete definition

Nothing

Methods

askTracerIO :: m Tracer

default askTracerIO :: (MonadTrans t, MonadTracingIO n, m ~ t n) => m Tracer

data TracerProvider

Constructors

TracerProvider

Fields

tracerProviderGetTracer :: InstrumentationScope -> IO Tracer

tracerProviderShutdown :: IO ()

tracerProviderForceFlush :: IO ()

data Tracer

Constructors

Tracer

Fields

tracerInstrumentationScope :: InstrumentationScope

tracerNow :: IO Timestamp

tracerStartSpan :: CallStack -> Context -> SpanSpec -> IO (MutableSpan, [Pair])

tracerProcessSpan :: Span Attrs -> IO ()

tracerSpanAttrsLimits :: AttrsLimits 'AttrsForSpan

tracerSpanEventAttrsLimits :: AttrsLimits 'AttrsForSpanEvent

tracerSpanLinkAttrsLimits :: AttrsLimits 'AttrsForSpanLink

contextBackendSpan :: ContextBackend MutableSpan

contextKeySpan :: ContextKey MutableSpan

data SpanContext

Constructors

SpanContext

Fields

spanContextTraceId :: TraceId

spanContextSpanId :: SpanId

spanContextTraceFlags :: TraceFlags

spanContextTraceState :: TraceState

spanContextIsRemote :: Bool

emptySpanContext :: SpanContext

spanContextIsValid :: SpanContext -> Bool

spanContextIsSampled :: SpanContext -> Bool

data TraceId

Constructors

TraceId

Fields

traceIdHi :: Word64

traceIdLo :: Word64

traceIdToHexText :: TraceId -> Text

traceIdToBytesVector :: TraceId -> Vector Word8

traceIdToHexBuilder :: TraceId -> Builder

traceIdToBytesBuilder :: TraceId -> Builder

emptyTraceId :: TraceId

traceIdFromWords :: Word64 -> Word64 -> TraceId

newtype SpanId

Constructors

SpanId

Fields

spanIdLo :: Word64

spanIdToHexText :: SpanId -> Text

spanIdToBytesVector :: SpanId -> Vector Word8

spanIdToHexBuilder :: SpanId -> Builder

spanIdToBytesBuilder :: SpanId -> Builder

emptySpanId :: SpanId

spanIdFromWords :: Word64 -> SpanId

newtype TraceFlags

Constructors

TraceFlags

Fields

unTraceFlags :: Word8

traceFlagsToHexText :: TraceFlags -> Text

traceFlagsToHexBuilder :: TraceFlags -> Builder

traceFlagsSampled :: TraceFlags

isSampledFlagSet :: TraceFlags -> Bool

newtype TraceState

Constructors

TraceState

Fields

unTraceState :: HashMap Text Text

emptyTraceState :: TraceState

nullTraceState :: TraceState -> Bool

sizeTraceState :: TraceState -> Int

memberTraceState :: Key Text -> TraceState -> Bool

lookupTraceState :: Key Text -> TraceState -> Maybe Text

findWithDefaultTraceState :: Text -> Key Text -> TraceState -> Text

deleteTraceState :: Key Text -> TraceState -> TraceState

filterTraceState :: (Text -> Bool) -> TraceState -> TraceState

filterWithKeyTraceState :: (Key Text -> Text -> Bool) -> TraceState -> TraceState

foldMapWithKeyTraceState :: forall m. Monoid m => (Key Text -> Text -> m) -> TraceState -> m

toListTraceState :: TraceState -> [(Key Text, Text)]

newtype TraceStateBuilder a

Constructors

TraceStateBuilder

Fields

unTraceStateBuilder :: Either (DList TraceStateError) a

buildTraceState :: forall m. MonadThrow m => TraceStateBuilder TraceState -> m TraceState

buildTraceStatePure :: TraceStateBuilder TraceState -> Either TraceStateErrors TraceState

newtype SpanEvents (attrs :: AttrsFor -> Type)

Constructors

SpanEvents

Fields

unSpanEvents :: DList (SpanEvent attrs)

spanEventsFromList :: [SpanEvent attrs] -> SpanEvents attrs

spanEventsToList :: SpanEvents attrs -> [SpanEvent attrs]

freezeAllSpanEventAttrs :: AttrsLimits 'AttrsForSpanEvent -> SpanEvents AttrsBuilder -> SpanEvents Attrs

data SpanEvent (attrs :: AttrsFor -> Type)

Constructors

SpanEvent

Fields

spanEventName :: SpanEventName

spanEventTimestamp :: Timestamp

spanEventAttrs :: attrs 'AttrsForSpanEvent

freezeSpanEventAttrs :: AttrsLimits 'AttrsForSpanEvent -> SpanEvent AttrsBuilder -> SpanEvent Attrs

newtype SpanEventSpecs

Constructors

SpanEventSpecs

Fields

unSpanEventSpecs :: DList SpanEventSpec

singletonSpanEventSpecs :: SpanEventSpec -> SpanEventSpecs

spanEventSpecsFromList :: [SpanEventSpec] -> SpanEventSpecs

spanEventSpecsToList :: SpanEventSpecs -> [SpanEventSpec]

data SpanEventSpec

Constructors

SpanEventSpec

Fields

spanEventSpecName :: SpanEventName

spanEventSpecTimestamp :: TimestampSource

spanEventSpecAttrs :: AttrsBuilder 'AttrsForSpanEvent

defaultSpanEventSpec :: SpanEventSpec

newtype SpanEventName

Constructors

SpanEventName

Fields

unSpanEventName :: Text

newtype SpanLinks (attrs :: AttrsFor -> Type)

Constructors

SpanLinks

Fields

unSpanLinks :: DList (SpanLink attrs)

spanLinksFromList :: [SpanLink attrs] -> SpanLinks attrs

spanLinksToList :: SpanLinks attrs -> [SpanLink attrs]

freezeAllSpanLinkAttrs :: AttrsLimits 'AttrsForSpanLink -> SpanLinks AttrsBuilder -> SpanLinks Attrs

newtype SpanLinkSpecs

Constructors

SpanLinkSpecs

Fields

unSpanLinkSpecs :: DList SpanLinkSpec

singletonSpanLinkSpecs :: SpanLinkSpec -> SpanLinkSpecs

spanLinkSpecsFromList :: [SpanLinkSpec] -> SpanLinkSpecs

spanLinkSpecsToList :: SpanLinkSpecs -> [SpanLinkSpec]

data SpanLink (attrs :: AttrsFor -> Type)

Constructors

SpanLink

Fields

spanLinkSpanContext :: SpanContext

spanLinkAttrs :: attrs 'AttrsForSpanLink

freezeSpanLinkAttrs :: AttrsLimits 'AttrsForSpanLink -> SpanLink AttrsBuilder -> SpanLink Attrs

newtype SpanLinkName

Constructors

SpanLinkName

Fields

unSpanLinkName :: Text

data SpanLinkSpec

Constructors

SpanLinkSpec

Fields

spanLinkSpecSpanContext :: SpanContext

spanLinkSpecAttrs :: AttrsBuilder 'AttrsForSpanLink

defaultSpanLinkSpec :: SpanLinkSpec

data SpanSpec

Constructors

SpanSpec

Fields

spanSpecName :: SpanName

spanSpecParentContext :: Maybe Context

spanSpecStart :: TimestampSource

spanSpecKind :: SpanKind

spanSpecAttrs :: AttrsBuilder 'AttrsForSpan

spanSpecLinks :: SpanLinkSpecs

defaultSpanSpec :: SpanSpec

data UpdateSpanSpec

Constructors

UpdateSpanSpec

Fields

updateSpanSpecName :: Maybe SpanName

updateSpanSpecStatus :: Maybe SpanStatus

updateSpanSpecAttrs :: Maybe (AttrsBuilder 'AttrsForSpan)

updateSpanSpecEvents :: Maybe SpanEventSpecs

defaultUpdateSpanSpec :: UpdateSpanSpec

buildSpanUpdater :: forall m. Monad m => m Timestamp -> UpdateSpanSpec -> m (Span AttrsBuilder -> Span AttrsBuilder)

recordException :: SomeException -> Bool -> TimestampSource -> AttrsBuilder 'AttrsForSpanEvent -> UpdateSpanSpec

exceptionEvent :: SomeException -> Bool -> TimestampSource -> AttrsBuilder 'AttrsForSpanEvent -> SpanEventSpec

newtype SpanName

Constructors

SpanName

Fields

unSpanName :: Text

newtype MutableSpan

Constructors

MutableSpan

Fields

unMutableSpan :: IORef (Span AttrsBuilder)

unsafeNewMutableSpan :: Span AttrsBuilder -> IO MutableSpan

unsafeReadMutableSpan :: MutableSpan -> IO (Span AttrsBuilder)

unsafeModifyMutableSpan :: MutableSpan -> (Span AttrsBuilder -> (Span AttrsBuilder, a)) -> IO a

data Span (attrs :: AttrsFor -> Type)

Constructors

Span

Fields

spanLineage :: SpanLineage

spanContext :: SpanContext

spanName :: SpanName

spanStatus :: SpanStatus

spanStart :: Timestamp

spanFrozenAt :: SpanFrozenAt attrs

spanKind :: SpanKind

spanAttrs :: attrs 'AttrsForSpan

spanLinks :: SpanLinks attrs

spanEvents :: SpanEvents attrs

spanIsRecording :: Bool

spanInstrumentationScope :: InstrumentationScope

spanIsRemote :: Span attrs -> Bool

spanIsSampled :: Span attrs -> Bool

spanIsRoot :: Span attrs -> Bool

spanIsChildOf :: Span attrs -> Span attrs -> Bool

type family SpanFrozenAt (attrs :: AttrsFor -> Type) :: Type where ...

Equations

SpanFrozenAt AttrsBuilder = Maybe Timestamp

SpanFrozenAt Attrs = SpanFrozenTimestamp

data SpanFrozenTimestamp

Constructors

SpanFrozenTimestampFrozen Timestamp

SpanFrozenTimestampEnded Timestamp

frozenTimestamp :: SpanFrozenTimestamp -> Timestamp

freezeSpan :: Timestamp -> AttrsLimits 'AttrsForSpanLink -> AttrsLimits 'AttrsForSpanEvent -> AttrsLimits 'AttrsForSpan -> Span AttrsBuilder -> Span Attrs

data SpanLineage

Constructors

SpanLineageRoot

SpanLineageChildOf SpanContext

data SpanKind

Constructors

SpanKindServer

SpanKindClient

SpanKindProducer

SpanKindConsumer

SpanKindInternal

data SpanStatus

Constructors

SpanStatusUnset

SpanStatusError Text

SpanStatusOk